+++ box2d/Source/Collision/b2Distance.cpp
@@ -31,8 +31,8 @@ static int32 ProcessTwo(b2Vec2* p1Out, b2Vec2* p2Out, b2Vec2* p1s, b2Vec2* p2s, 
 	// If in point[1] region
 	b2Vec2 r = -points[1];
 	b2Vec2 d = points[0] - points[1];
-	float32 length = d.Normalize();
-	float32 lambda = b2Dot(r, d);
+	float64 length = d.Normalize();
+	float64 lambda = b2Dot(r, d);
 	if (lambda <= 0.0f || length < FLT_EPSILON)
 	{
 		// The simplex is reduced to a point.
@@ -66,9 +66,9 @@ static int32 ProcessThree(b2Vec2* p1Out, b2Vec2* p2Out, b2Vec2* p1s, b2Vec2* p2s
 	b2Vec2 ac = c - a;
 	b2Vec2 bc = c - b;
 
-	float32 sn = -b2Dot(a, ab), sd = b2Dot(b, ab);
-	float32 tn = -b2Dot(a, ac), td = b2Dot(c, ac);
-	float32 un = -b2Dot(b, bc), ud = b2Dot(c, bc);
+	float64 sn = -b2Dot(a, ab), sd = b2Dot(b, ab);
+	float64 tn = -b2Dot(a, ac), td = b2Dot(c, ac);
+	float64 un = -b2Dot(b, bc), ud = b2Dot(c, bc);
 
 	// In vertex c region?
 	if (td <= 0.0f && ud <= 0.0f)
@@ -88,18 +88,18 @@ static int32 ProcessThree(b2Vec2* p1Out, b2Vec2* p2Out, b2Vec2* p1s, b2Vec2* p2s
 	b2Assert(sn > 0.0f || tn > 0.0f);
 	b2Assert(sd > 0.0f || un > 0.0f);
 
-	float32 n = b2Cross(ab, ac);
+	float64 n = b2Cross(ab, ac);
 
 	// Should not be in edge ab region.
-	float32 vc = n * b2Cross(a, b);
+	float64 vc = n * b2Cross(a, b);
 	b2Assert(vc > 0.0f || sn > 0.0f || sd > 0.0f);
 
 	// In edge bc region?
-	float32 va = n * b2Cross(b, c);
+	float64 va = n * b2Cross(b, c);
 	if (va <= 0.0f && un >= 0.0f && ud >= 0.0f)
 	{
 		b2Assert(un + ud > 0.0f);
-		float32 lambda = un / (un + ud);
+		float64 lambda = un / (un + ud);
 		*p1Out = p1s[1] + lambda * (p1s[2] - p1s[1]);
 		*p2Out = p2s[1] + lambda * (p2s[2] - p2s[1]);
 		p1s[0] = p1s[2];
@@ -109,11 +109,11 @@ static int32 ProcessThree(b2Vec2* p1Out, b2Vec2* p2Out, b2Vec2* p1s, b2Vec2* p2s
 	}
 
 	// In edge ac region?
-	float32 vb = n * b2Cross(c, a);
+	float64 vb = n * b2Cross(c, a);
 	if (vb <= 0.0f && tn >= 0.0f && td >= 0.0f)
 	{
 		b2Assert(tn + td > 0.0f);
-		float32 lambda = tn / (tn + td);
+		float64 lambda = tn / (tn + td);
 		*p1Out = p1s[0] + lambda * (p1s[2] - p1s[0]);
 		*p2Out = p2s[0] + lambda * (p2s[2] - p2s[0]);
 		p1s[1] = p1s[2];
@@ -123,12 +123,12 @@ static int32 ProcessThree(b2Vec2* p1Out, b2Vec2* p2Out, b2Vec2* p1s, b2Vec2* p2s
 	}
 
 	// Inside the triangle, compute barycentric coordinates
-	float32 denom = va + vb + vc;
+	float64 denom = va + vb + vc;
 	b2Assert(denom > 0.0f);
 	denom = 1.0f / denom;
-	float32 u = va * denom;
-	float32 v = vb * denom;
-	float32 w = 1.0f - u - v;
+	float64 u = va * denom;
+	float64 v = vb * denom;
+	float64 w = 1.0f - u - v;
 	*p1Out = u * p1s[0] + v * p1s[1] + w * p1s[2];
 	*p2Out = u * p2s[0] + v * p2s[1] + w * p2s[2];
 	return 3;
@@ -147,7 +147,7 @@ static bool InPoints(const b2Vec2& w, const b2Vec2* points, int32 pointCount)
 	return false;
 }
 
-float32 b2Distance(b2Vec2* p1Out, b2Vec2* p2Out, const b2Shape* shape1, const b2Shape* shape2)
+float64 b2Distance(b2Vec2* p1Out, b2Vec2* p2Out, const b2Shape* shape1, const b2Shape* shape2)
 {
 	b2Vec2 p1s[3], p2s[3];
 	b2Vec2 points[3];
@@ -156,7 +156,7 @@ float32 b2Distance(b2Vec2* p1Out, b2Vec2* p2Out, const b2Shape* shape1, const b2
 	*p1Out = shape1->m_position;
 	*p2Out = shape2->m_position;
 
-	float32 vSqr = 0.0f;
+	float64 vSqr = 0.0f;
 	const int32 maxIterations = 20;
 	for (int32 iter = 0; iter < maxIterations; ++iter)
 	{
@@ -166,7 +166,7 @@ float32 b2Distance(b2Vec2* p1Out, b2Vec2* p2Out, const b2Shape* shape1, const b2
 
 		vSqr = b2Dot(v, v);
 		b2Vec2 w = w2 - w1;
-		float32 vw = b2Dot(v, w);
+		float64 vw = b2Dot(v, w);
 		if (vSqr - vw <= 0.01f * vSqr || InPoints(w, points, pointCount)) // or w in points
 		{
 			if (pointCount == 0)
@@ -211,7 +211,7 @@ float32 b2Distance(b2Vec2* p1Out, b2Vec2* p2Out, const b2Shape* shape1, const b2
 			return 0.0f;
 		}
 
-		float32 maxSqr = -FLT_MAX;
+		float64 maxSqr = -FLT_MAX;
 		for (int32 i = 0; i < pointCount; ++i)
 		{
 			maxSqr = b2Max(maxSqr, b2Dot(points[i], points[i]));
